Add self option to filter options
